requirements.txt
streamlit
numpy
pillow
torch
transformers
scikit-learn
